nesting of files; unnecessary styles; add fbg2; img height;
