+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
-import Header from "./components/header/Header";
-import Main from "./components/main/Main";
-import Footer from "./components/footer/Footer";
+import Header from "./Header";
+import Main from "./Main";
+import Footer from "./Footer";
 
 class App extends Component {
   render() {
