+++ src/Header/index.js
@@ -1,9 +1,9 @@
 import React, { Component } from "react";
 import styled from "styled-components";
 import { Wrapper } from "../defaults";
-import Logo from "../logo/Logo";
-import Search from "../search/Search";
-import Nav from "../nav/Nav";
+import Logo from "./Logo";
+import Search from "./Search";
+import Nav from "./Nav";
 
 const HeaderBox = styled.header`
   box-shadow: 0px 0.5px 0px rgba(72, 72, 72, 0.3);
