+++ src/index.js
@@ -1,7 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
 import "normalize.css";
+import "flexboxgrid2";
+import "./index.css";
 import App from "./App";
 import registerServiceWorker from "./registerServiceWorker";
